refactor(suite-native): crypto to crypto conversion formatter hook
